test: clean up SelectorCache tests
